+++ examples/resources/dokploy_backup/import.sh
@@ -1,0 +1,1 @@
+terraform import dokploy_backup.db_nightly b1a2b3c4d5e6f7g8h9i0j